refactor(dataset): parser renamed to dataset
